requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
psycopg2-binary==2.9.10
httpx==0.28.1
pytest==8.3.4
